add toggle menu/cart

+++ app/Http/Controllers/Admin/FoodController.php
@@ -142,9 +142,9 @@ class FoodController extends Controller
     public function destroy(Request $request)
     {
         $food = Food::find($request->id);
-        // if( $food->user_id != Auth::id() ) {
-        //     abort("403");
-        // }
+        if( $food->user_id != Auth::id() ) {
+            abort("403");
+        }
 
         Storage::delete($food->image);
         if (!is_null($food)) {
